# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Premium Matrimony API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterBody {
  email: string;
  name: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface Profile {
  id: number;
  name: string;
  age: number;
  gender: string;
  religion: string;
  /** @nullable */
  caste?: string | null;
  city: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  profession?: string | null;
  photo: string;
  bio: string;
  featured: boolean;
  createdAt?: string;
}

export interface ProfileInput {
  name: string;
  age: number;
  gender: string;
  religion: string;
  caste?: string;
  city: string;
  state?: string;
  education?: string;
  profession?: string;
  photo?: string;
  bio: string;
}

export interface Interest {
  id: number;
  fromProfileId: number;
  toProfileId: number;
  status: string;
  createdAt: string;
}

export interface InterestInput {
  fromProfileId: number;
  toProfileId: number;
}

export interface Story {
  id: number;
  coupleName: string;
  story: string;
  photo: string;
  marriageYear: number;
  /** @nullable */
  city?: string | null;
}

export interface StoryInput {
  coupleName: string;
  story: string;
  photo?: string;
  marriageYear: number;
  city?: string;
}

export interface PlatformStats {
  totalMembers: number;
  totalMatches: number;
  successStories: number;
  citiesCovered: number;
}

export type ListProfilesParams = {
gender?: string;
religion?: string;
minAge?: number;
maxAge?: number;
city?: string;
limit?: number;
offset?: number;
};