# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Premium Matrimony API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  "email": zod.string(),
  "name": zod.string(),
  "password": zod.string()
})


/**
 * @summary Log in
 */
export const LoginBody = zod.object({
  "email": zod.string(),
  "password": zod.string()
})

export const LoginResponse = zod.object({
  "id": zod.number(),
  "email": zod.string(),
  "name": zod.string(),
  "createdAt": zod.string()
})


/**
 * @summary Log out
 */
export const LogoutResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  "id": zod.number(),
  "email": zod.string(),
  "name": zod.string(),
  "createdAt": zod.string()
})


/**
 * @summary List profiles with filters
 */
export const ListProfilesQueryParams = zod.object({
  "gender": zod.coerce.string().optional(),
  "religion": zod.coerce.string().optional(),
  "minAge": zod.coerce.number().optional(),
  "maxAge": zod.coerce.number().optional(),
  "city": zod.coerce.string().optional(),
  "limit": zod.coerce.number().optional(),
  "offset": zod.coerce.number().optional()
})

export const ListProfilesResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "age": zod.number(),
  "gender": zod.string(),
  "religion": zod.string(),
  "caste": zod.string().nullish(),
  "city": zod.string(),
  "state": zod.string().nullish(),
  "education": zod.string().nullish(),
  "profession": zod.string().nullish(),
  "photo": zod.string(),
  "bio": zod.string(),
  "featured": zod.boolean(),
  "createdAt": zod.string().optional()
})
export const ListProfilesResponse = zod.array(ListProfilesResponseItem)


/**
 * @summary Create a new profile
 */
export const CreateProfileBody = zod.object({
  "name": zod.string(),
  "age": zod.number(),
  "gender": zod.string(),
  "religion": zod.string(),
  "caste": zod.string().optional(),
  "city": zod.string(),
  "state": zod.string().optional(),
  "education": zod.string().optional(),
  "profession": zod.string().optional(),
  "photo": zod.string().optional(),
  "bio": zod.string()
})


/**
 * @summary Get featured profiles
 */
export const ListFeaturedProfilesResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "age": zod.number(),
  "gender": zod.string(),
  "religion": zod.string(),
  "caste": zod.string().nullish(),
  "city": zod.string(),
  "state": zod.string().nullish(),
  "education": zod.string().nullish(),
  "profession": zod.string().nullish(),
  "photo": zod.string(),
  "bio": zod.string(),
  "featured": zod.boolean(),
  "createdAt": zod.string().optional()
})
export const ListFeaturedProfilesResponse = zod.array(ListFeaturedProfilesResponseItem)


/**
 * @summary Get a single profile
 */
export const GetProfileParams = zod.object({
  "id": zod.coerce.number()
})

export const GetProfileResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "age": zod.number(),
  "gender": zod.string(),
  "religion": zod.string(),
  "caste": zod.string().nullish(),
  "city": zod.string(),
  "state": zod.string().nullish(),
  "education": zod.string().nullish(),
  "profession": zod.string().nullish(),
  "photo": zod.string(),
  "bio": zod.string(),
  "featured": zod.boolean(),
  "createdAt": zod.string().optional()
})


/**
 * @summary List interests
 */
export const ListInterestsResponseItem = zod.object({
  "id": zod.number(),
  "fromProfileId": zod.number(),
  "toProfileId": zod.number(),
  "status": zod.string(),
  "createdAt": zod.string()
})
export const ListInterestsResponse = zod.array(ListInterestsResponseItem)


/**
 * @summary Send interest to a profile
 */
export const SendInterestBody = zod.object({
  "fromProfileId": zod.number(),
  "toProfileId": zod.number()
})


/**
 * @summary List success stories
 */
export const ListStoriesResponseItem = zod.object({
  "id": zod.number(),
  "coupleName": zod.string(),
  "story": zod.string(),
  "photo": zod.string(),
  "marriageYear": zod.number(),
  "city": zod.string().nullish()
})
export const ListStoriesResponse = zod.array(ListStoriesResponseItem)


/**
 * @summary Submit a success story
 */
export const CreateStoryBody = zod.object({
  "coupleName": zod.string(),
  "story": zod.string(),
  "photo": zod.string().optional(),
  "marriageYear": zod.number(),
  "city": zod.string().optional()
})


/**
 * @summary Get platform statistics
 */
export const GetStatsResponse = zod.object({
  "totalMembers": zod.number(),
  "totalMatches": zod.number(),
  "successStories": zod.number(),
  "citiesCovered": zod.number()
})